app/requirements.txt
openai>=1.55,<2
faiss-cpu==1.8.0.post1
numpy>=1.26,<2.0
gradio>=4.44,<5
huggingface_hub>=0.23,<1.0
starlette<1.0.0
fastapi<0.116
requests>=2.32
tqdm>=4.66
